Move prepare_buffers to prepare_buffer.hpp (API Change)
fix #338
