+++ CHANGELOG.md
@@ -6,6 +6,7 @@ API Changes
 
 * Return http::error::end_of_stream on HTTP read eof
 * Remove placeholders
+* Move prepare_buffers to prepare_buffer.hpp
 
 --------------------------------------------------------------------------------
