+++ test/http/design.cpp
@@ -6,7 +6,7 @@
 //
 
 #include <beast/core/flat_buffer.hpp>
-#include <beast/core/prepare_buffers.hpp>
+#include <beast/core/prepare_buffer.hpp>
 #include <beast/http/chunk_encode.hpp>
 #include <beast/http/read.hpp>
 #include <beast/http/write.hpp>
